Refactor code to add BookL1Struct class and update benchmark.py
This commit refactors the code to add a new class called BookL1Struct in the utils.py module. The BookL1Struct class is a subclass of the msgspec.Struct class and represents a structured book level 1 data. The benchmark.py module is also updated to include a new benchmark for creating instances of the BookL1Struct class.